DeepCode/flask_app/.ipynb_checkpoints/requirements-checkpoint.txt
flask
python-dotenv
google-genai
google-generativeai
requests
pytest
pytest-mock
